Decimal::divide_integer: cast ~FLAG_* to uint8_t for -Werror=sign-conversion
The unit-divisor fast path added in c069942 used '&= ~FLAG_SIGN' and
'&= ~FLAG_INTEGER_LITERAL' directly. Because ~ on a std::uint8_t
promotes to int, the result is a negative int (~0x01 = -2, ~0x40 = -65
under two's complement int semantics). Assigning that back to
flags_ (std::uint8_t) triggers GCC's -Werror=sign-conversion in the
CI configuration (SOURCEMETA_CORE_TESTS + CMAKE_COMPILE_WARNING_AS_ERROR).
Clang without that warning did not catch it locally.

Fix: wrap each ~FLAG_* in static_cast<std::uint8_t>(...) so the
narrowing is explicit. Behavior identical; only quiets the warning.

Verified locally with make configure (which sets
-DCMAKE_COMPILE_WARNING_AS_ERROR:BOOL=ON) and full builds of both
sourcemeta_core_numeric and the numeric_unit + ibm_dectest_unit test
targets; 776/776 numeric tests pass and 8701/59-skip/0-fail on the
IBM dectest suite (baseline maintained).

+++ src/lang/numeric/decimal.cc
@@ -1306,14 +1306,14 @@ auto Decimal::divide_integer(const Decimal &other) const -> Decimal {
   if (this->exponent_ >= 0) {
     const Decimal one{1};
     Decimal other_abs = other;
-    other_abs.flags_ &= ~FLAG_SIGN;
+    other_abs.flags_ &= static_cast<std::uint8_t>(~FLAG_SIGN);
     if (other_abs == one) {
       Decimal result = *this;
-      result.flags_ &= ~FLAG_INTEGER_LITERAL;
+      result.flags_ &= static_cast<std::uint8_t>(~FLAG_INTEGER_LITERAL);
       if (result_negative) {
         result.flags_ |= FLAG_SIGN;
       } else {
-        result.flags_ &= ~FLAG_SIGN;
+        result.flags_ &= static_cast<std::uint8_t>(~FLAG_SIGN);
       }
       return result;
     }
